Added temp directory option to Configuration
The option was there already, but undocumented -- now,
setting config.temp to a directory will cause that directory to be
created prior to starting the run, and the output of each job will
be stored in the directory (as shortname.machine). If config.temp is
false, no directory is created and no output is saved.